At this point everything works fine except Document generating
the generation and manipulation of documents works for AC and AV but the problem is the content overflows onto each others

+++ OTAS-Backend/DTO/Get/StatsRequestsDTO.cs
@@ -6,7 +6,7 @@
 
         public int AllOngoingRequestsCount { get; set; } = 0;
 
-        public int? HoursPassedSinceLastRequest { get; set; }
+        public decimal? HoursPassedSinceLastRequest { get; set; }
 
         public int OrdreMissionsAllTimeCount { get; set; } = 0;
